requirements.txt
discord.py
python-dotenv
flask
requests
beautifulsoup4
lxml
